09 15 2025
Had a great time working or starting to work on this new problem.

I have learned that the JOIN order does not matter on INNER joins!

Amazing.  I am sure I had learned this previously, but I need to reinforce it over and over again I guess.

So, the next time I work on this I will probably add the population column to the result set.

It seems like I am missing something but I will slowly get to this.